refactor(prompt-generator): move to core/prompt-generator
