+++ packages/core/src/prompt-generator/about-this.ts
