+++ packages/core/src/prompt-generator/action.ts
