+++ packages/core/src/prompt-generator/index.ts
@@ -1,0 +1,1 @@
+export * from './prompt';
