--- packages/genui-renderer/src/server/prompt/index.ts
@@ -1,1 +1,0 @@
-export * from './prompt';
