+++ sites/playground/server/package.json
@@ -18,6 +18,7 @@
     "@modelcontextprotocol/server-sequential-thinking": "^2025.7.1",
     "@openrouter/ai-sdk-provider": "^1.1.2",
     "@opentiny/genui-sdk": "workspace:*",
+    "@opentiny/genui-sdk-core": "workspace:*",
     "ai": "5.0.10",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
