+++ sites/playground/server/src/chat-genui.ts
@@ -5,7 +5,8 @@ import fs from 'node:fs/promises';
 import path from 'node:path';
 import { z } from 'zod';
 import { fileURLToPath } from 'node:url';
-import { rendererConfig, ngRendererConfig, genPrompt } from '@opentiny/genui-sdk/server';
+import { rendererConfig, ngRendererConfig } from '@opentiny/genui-sdk/server';
+import { genPrompt } from '@opentiny/genui-sdk-core';
 import { Client } from '@modelcontextprotocol/sdk/client/index.js';
 import { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp.js';
 import { SSEClientTransport } from '@modelcontextprotocol/sdk/client/sse.js';
